refactor: get rid of adpater file and place the logic inside codegen-head
